Revert if conditions

+++ lumina_next_t2i_mini/train_vae_swap_layers.py
@@ -231,9 +231,10 @@ def main(args):
             [cv2.resize(numpy.array(frame), (512, 512)) for frame in images])  # Resize all frames
         #  batch_size, num_channels, num_frames, height, width = x.shape
         frames_tensor = torch.from_numpy(frames_resized).permute(3, 0, 1, 2).unsqueeze(0)
-        frames_tensor = frames_tensor.to(dtype=torch.float32, non_blocking=True) / 127.5 - 1  # Normalize
+        frames_tensor = frames_tensor.to(dtype=torch.float16, non_blocking=True) / 127.5 - 1  # Normalize
 
         with torch.no_grad():
+            frames_tensor = frames_tensor.to(dtype=torch.float32)
             latent = vae.encode(frames_tensor).latent_dist.sample().mul_(vae_scale)
             assert not torch.isnan(latent).any(), "NaN detected in latent!"
             latent = latent.to(device)
